change .js to .jsx
