+++ .eslintrc.json
@@ -1,6 +1,5 @@
 {
   "extends": ["next", "next/core-web-vitals", "prettier", "eslint:recommended"],
-  "ignorePatterns": ["*.config.js"],
   "plugins": ["prettier", "react"],
   "env": {
     "browser": true
