+++ components/Atoms/button/ContBlue.jsx
@@ -1,28 +1,27 @@
-import { ChakraProvider, Box } from "@chakra-ui/react";
+import { Box } from "@chakra-ui/react";
 
-export function OutOrange(text) {
+export function ContBlue(text) {
   return (
-    <ChakraProvider>
       <Box
+        type="submit"
         as="button"
         height="40px"
         lineHeight="1.2"
         transition="all 0.2s cubic-bezier(.08,.52,.52,1)"
         border="1px"
-        px="20px"
-        width="100px"
+        px="30px"
         borderRadius="8px"
         fontSize="14px"
         fontWeight="semibold"
-        bg="none"
-        borderColor="orange"
-        color="orange"
+        bg="#062572"
+        borderColor="#062572"
+        color="white"
         _hover={{
-          opacity: 0.5,
+          bg: "#062572",
+          color: "white",
         }}
       >
         {text}
       </Box>
-    </ChakraProvider>
   );
 }
