+++ components/Atoms/button/OutWhite.jsx
@@ -1,8 +1,7 @@
-import { ChakraProvider, Box } from "@chakra-ui/react";
+import { Box } from "@chakra-ui/react";
 
-export function OutBlue(text) {
+export function OutWhite(text) {
   return (
-    <ChakraProvider>
       <Box
         as="button"
         height="40px"
@@ -10,19 +9,18 @@ export function OutBlue(text) {
         transition="all 0.2s cubic-bezier(.08,.52,.52,1)"
         border="1px"
         px="20px"
+        width="100px"
         borderRadius="8px"
         fontSize="14px"
         fontWeight="semibold"
         bg="none"
-        borderColor="#062572"
-        color="#062572"
+        borderColor="white"
+        color="white"
         _hover={{
-          bg: "#062572",
-          color: "white",
+          opacity: 0.5,
         }}
       >
         {text}
       </Box>
-    </ChakraProvider>
   );
 }
