+++ pages/contact.jsx
@@ -4,7 +4,7 @@ import { useRouter } from "next/router";
 import * as gtag from "@/src/gtag";
 import React, { useState } from "react";
 
-import {TempContact} from "@/Templates/contact";
+import {TempContact} from "@/components/Templates/contact";
 
 export default function Contact() {
   const registerUser = async (event) => {
